Update expanding_toolbar.py
This file was modified by adding three lines of code (lines 10, line 69, line 70) to solve Issue #183 . Line 10 just imports time and line 69 is a comment. By putting the time.sleep(0.25) command in line 70, there is a 0.25 second delay before an expanded widget "contracts" in the event that the hovering mouse accidentally moves from the widget area.