Fix package metadata (#409)
This commit fixes serval issues with the package metadata. The first is
the trove classifiers were incorrectly listing the package as supporting
multiple versions of python, however the current metal repo only
supports 3.7 (and a narrow subset too) so the unsupported python
versions are removed. Additionally, it was listing metal as a console
application which it is not, it's a GUI and an API. Then additional
metadata is added to add links to the tracker, documentation, and source
repo.