Move draw_sample_holder to render_chips
Previously it was done in render_chip, which still makes sense if you
only have a design with one chip.

When your design involves two or more chips, then it makes sense to draw
the sample holder only once. As such, the appropriate place is not in
render_chip, but in render_chips.

The x-y size of the vacuum box should be able to host all of the chips.

Current implementation still draws the information on sample holder
top/bottom from each chip.
I will fix this in future implementation by moving the information to
the design level