Move sampleholder size extraction to design level
For a design containing more than one chip, it makes sense that the
sampleholder is a property of the design and not each chip. Otherwise it
creates the impression that there is a specific sampleholder object for
each chip.

The try except clause is there to make sure that it is still compatible
with DesignPlanar and other QDesign classes that encode the sample
holder size information at the chip level.