Disable chip_not_main func
Latest rebase introduces additional function that essentially checks if
your chip name is not 'main'. If it's not, then it stops ansys from
rendering. It seems that the metal team has their own plan with this, so
while waiting for their update, I'm going to temporarily comment out
these lines so that I can continue working on the 3d stack work.