Render all chips by default for FlipChip case
Previous the self.get_all_chips() only returns chip name containing
qcomponents that are to be drawn. While this is not really a problem for
the planar case, it is for the flipchip case.

In the case of flipchip, even if we request for qubit to be drawn on the
ansys model, we need to render the C_chip as well. This is not handled
very well by the get_all_chips() function, i.e. it only draws the Q_chip.
There are few possible ways to fix this, but I eventually decided for a
quick fix by overwriting the chip_list whenever it identifies a FlipChip
device design. Another alternative involves me having to modify
more classes, which is not ideal at this point in time.

I hope that this gets a better fix in the future. It should be the case
that even though no qcomponents are registered for that particular chip
in the table, it should still be rendered in ANSYS or at least we should
have the option to render the chip.