Flip-Chip stack (#677)
* Added a new Qdesign class called DesignFlipChip

* DesignFlipChip is registered in the main qiskit_metal module

* fix axis problem for draw_rect_corner when non-zero qc_chip_z is used

* junction can be drawn for non zero chip_z

* transmonpocket can be used for different chip name

* Replaced z position with its unparsed version, so that it only gets parsed once. Previous usage of qc_chip_z for the z position means that it gets parsed twice, and this leads to unit inconsistency

* Fix rendering of chip for non-zero center_z

render_chips:
Instead of determining cc_x, cc_y, cw_x, cw_y first and render each chip
(which does not work for a design containing more than one chip), the
code now defines cc_x,... as a dictionary with the chip_name as the key.

render_chip:
It now build the chip based on the corresponding cc_x, ...
Commented out draw_sample_holder in render_chip, the right place to draw_sample_holder is in render_chips.
This way, the code does not draw one sample holder for each chip

* Move draw_sample_holder to render_chips

Previously it was done in render_chip, which still makes sense if you
only have a design with one chip.

When your design involves two or more chips, then it makes sense to draw
the sample holder only once. As such, the appropriate place is not in
render_chip, but in render_chips.

The x-y size of the vacuum box should be able to host all of the chips.

Current implementation still draws the information on sample holder
top/bottom from each chip.
I will fix this in future implementation by moving the information to
the design level

* Move sample_holder_top, _bottom to design level

This is so that sample_holder_top and sample_holder_bottom is a property
of the design, instead of each chip. Such arrangement makes sense when
you consider a design with more than one chip.

To ensure continuity with the default dictionary, I maintain
sample_holder_top and sample_holder_bottom keys in each chip entry. But
it should be understood that future implementation would need to access
the sample_holder_top and _bottom information from self.variables.

Next commit will ensure that ansys_renderer will get the information
from self.variables.

* Remove chip_spacing from self.variables

We can access chip_spacing information from the center_z of each chip,
so this additional information is not needed anymore.

* Move sampleholder size extraction to design level

For a design containing more than one chip, it makes sense that the
sampleholder is a property of the design and not each chip. Otherwise it
creates the impression that there is a specific sampleholder object for
each chip.

The try except clause is there to make sure that it is still compatible
with DesignPlanar and other QDesign classes that encode the sample
holder size information at the chip level.

* Disable chip_not_main func

Latest rebase introduces additional function that essentially checks if
your chip name is not 'main'. If it's not, then it stops ansys from
rendering. It seems that the metal team has their own plan with this, so
while waiting for their update, I'm going to temporarily comment out
these lines so that I can continue working on the 3d stack work.

* Fix path render with non zero chip_z

When you use the sweep alonglength function in ansys, it seems that, the
sweep vector also has to be done at the right height.Previous
implementation forces the rendering to be done at z=0 all the time

* Fix pin render for chip with non-zero center_z

* Fix pin render on non-standard chip name

* Apply Black formatting

Files:
- design_flipchip
- ansys_renderer

* Apply yapf formatting

* Add chip option to transmon_cross dictionary

Previously, chip option was not specified by default, as a result, the
qcomponents are defined on the standard chip name called 'main'. We fix
this problem for the flipchip design class which requires us to specify
which chip it has be to be drawn on.

* Render all chips by default for FlipChip case

Previous the self.get_all_chips() only returns chip name containing
qcomponents that are to be drawn. While this is not really a problem for
the planar case, it is for the flipchip case.

In the case of flipchip, even if we request for qubit to be drawn on the
ansys model, we need to render the C_chip as well. This is not handled
very well by the get_all_chips() function, i.e. it only draws the Q_chip.
There are few possible ways to fix this, but I eventually decided for a
quick fix by overwriting the chip_list whenever it identifies a FlipChip
device design. Another alternative involves me having to modify
more classes, which is not ideal at this point in time.

I hope that this gets a better fix in the future. It should be the case
that even though no qcomponents are registered for that particular chip
in the table, it should still be rendered in ANSYS or at least we should
have the option to render the chip.

* Add a tutorial notebook for the FlipChip design

Content:
1. a tutorial notebook explaining the construction of the flipchip
   design, combined with an EPR analysis
2. a resonator component created for this tutorial.

* Tidy up documentations

The following items changed:

- The documentation for get_x_y_for_chip function in design_flipchip.py,
  the function seems to work fine with the flipchip structure.

- The example in the tutorial notebook: simplified the numbers, remove
  the output, corrected some errors.

- The resonator object built by readoutres_fc.py
	- it now inherits QComponent instead of BaseQubit class
	- the documentation for the function

* Apply yapf and improve code readability

Did some tidying to pass the CLI test
- Applied YAPF formatting to readoutres_fc.py
- Rename class and variables in readoutres_fc.py to improve linting
  score.
- Applied the necessary change to the flipchip tutorial notebook

* Tidy up transmon_cross.py

Applied the yapf format

Co-authored-by: Chris Warren <warrenc@chalmers.se>
Co-authored-by: Zlatko Minev <zminev@gmail.com>
Co-authored-by: Thomas G McConkey <53087709+ThomasGM4@users.noreply.github.com>
Co-authored-by: Priti Ashvin Shah <74020801+priti-ashvin-shah-ibm@users.noreply.github.com>