Create `get_gnd_qgeoms` function
The `get_gnd_qgeoms` function returns a list of qgeometry names associated with pins shorted to the ground plane.
It does this by looking at all pins in the design, and removing port pins which are a combination of not-shorted pins in `design.net_info()` and the `open_pins` parameter provided by the user.