Add MANIFEST.in (#895)
This adds a MANIFEST.in file so that some top-level text files make it
into the source distribution. requirements.txt and README.md are needed
in order to install the package from the source distribution because
they are accessed by setup.py. LICENSE.txt is also included because it
is best practice to include the license when distributing source code.