Shortened line for logging message
Previously, the logging message was long. To keep with the convention of approximately 100 characters per line, the line is broken down into multiple lines. The previous commit fixing this issue mistakenly had a comma at the end of each f-string which caused error. There is no need for comma at the end of each f-string when broken down into multiple lines.