Display option name and option value of failed runs in the Sweeper class' `run_sweep` method (#909)
* Display option name and option value of failed runs in the Sweeper class' `run_sweep` method

Fixes #907

Currently, when a run fails during the execution of Sweeper class' run_sweep method, the warning message, does not tell which option name or option value that failed.

It will be useful to add which option name + option value (key) combination thatdidn't execute properly for reference in case the user wantst to run the sweeperagain for the failed option values

This commit adds the option name and option value that failed to the warning message.

* Shorten line length of code changed

Previously, the code changed was too long. This commit breaks the line into multiple lines to keep with the convention of keeping the lines approximately 100 characters.