Display option name and option value of failed runs in the Sweeper class' run_sweep method (#913)
* Display option name and option value of failed runs in the Sweeper class' `run_sweep` method

Fixes #907

Currently, when a run fails during the execution of Sweeper class' run_sweep method, the warning message, does not tell which option name or option value that failed.

It will be useful to add which option name + option value (key) combination thatdidn't execute properly for reference in case the user wantst to run the sweeperagain for the failed option values

This commit adds the option name and option value that failed to the warning message.

* Shorten line length of code changed

Previously, the code changed was too long. This commit breaks the line into multiple lines to keep with the convention of keeping the lines approximately 100 characters.

* Revert "Shorten line length of code changed"

This reverts commit ac0a8ef70393a188a8542d73194c172d8a285aab.

The logging message does not work as expected when broken down into multiple lines. It throws an error when logging. Reverting this commit as it is not a workable solution.

* Shortened line for logging message

Previously, the logging message was long. To keep with the convention of approximately 100 characters per line, the line is broken down into multiple lines. The previous commit fixing this issue mistakenly had a comma at the end of each f-string which caused error. There is no need for comma at the end of each f-string when broken down into multiple lines.

---------

Co-authored-by: Priti Ashvin Shah <74020801+priti-ashvin-shah-ibm@users.noreply.github.com>