Expose result of convergence of LOM analysis' Q3D renderer to end user (#915)
* Extract convergence data from q3d simulation

Convergence data is extracted from q3d simulation. Implementation is
copied from Sweeping class

* Add is_converged attributed LumpedElementsSim

is_converged property is added to LumpedElementsSim class

* Add tests for getting convergence boolean

Add 2 tests to get convergence boolean.
1. Test to get True when simulation converge
2. Test to get False when simulation did not converge

* Add is_converged to data labels

is_converged variable is added to the data labels

* Add executed code illustrating is_converged key

The notebook is executed to illustrate the usage of is_converged
key. Example is shown in the notebook with accompanying textual
explanation.

* Add unexecuted notebook illustrating is_converged key

The notebook is updated to illustrate the use of is_converged
key. Code example accompanying textual explanation is added to the
notebook. The notebook is not executed.