GDS renderer: changed criterion for rounding paths
Method QGDSRenderer._qgeometry_to_gds() checks several conditions to
decide whether a shapely.geometry.LineString is drawn with rounded or
sharp corners. One of them takes into account that if the fillet radius
gets too small relative to the width of the line, the rounding cannot be
applied correctly anymore, hence drawing the line with sharp corners.

Before this commit, this was implemented via the criterion

  `qgeometry_element.fillet < qgeometry_element.width`;

however, this criterion is too strict because corners can well be
rounded also if 0.5*width < fillet radius < width. Therefore, this
commit changes this criterion to

  `qgeometry_element.fillet < 0.5 * qgeometry_element.width`.